Certificates of which *some* are required in order for the DMTU to be able to access UM-SSO-protected resources served from GMV's test environment (IP address = 5.9.173.37).
umsso.pem is for the IDP, and came in an email from Hector Hugo Perez Alonso to Tarek on 25 June 2013.
servercrt.pem is for the LDAP (so I don't know why it was sent to Magellium), and came in an email from Hector Hugo Perez Alonso to Tarek on 25 June 2013.
webs.novalocal.crt is the certificate (self-signed, IIRC, and not matching the domain from which it was served, IIRC) that was actually being served from 5.9.173.37 circa 26th June 2013.

Refer to ssl_certs.bat for a script that will build a truststore that encapsulates (only) these certificates.
TODO: We can whittle down the set of certs to just those that are actually required later. Also, need to think about a way of building and automatically integrating with a truststore whose contents depend on the target environment (e.g. Test versus Production).
TODO: Check that these certs aren't built into the DMTU's WAR file itself. Perhaps the certs within SVN need to live in a different folder...